feat: rename VM Configuration to Pool Status
- Renamed all "VM Configuration" references to "Pool Status"
- Renamed "VM Settings" to "Pool Settings"
- Renamed component directory from vm-config to pool-status
- Updated imports in dashboard and settings pages
- Changed "Set up VMs" to "Set up Pool" in UI

Fixes #871
